# src/StravaDiscordBot.DiscordApi/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "StravaDiscordBot.DiscordApi",
    "version": "1.0"
  },
  "paths": {
    "/v1/leaderboard/{leaderboardId}/participant/{participantId}/activity": {
      "get": {
        "tags": [
          "Activity"
        ],
        "operationId": "GetAllActivitiesForPeriod",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "leaderboardId",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "path",
            "name": "participantId",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "query",
            "name": "from",
            "type": "string",
            "format": "date-time"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/SummaryActivityResponse"
              }
            }
          }
        }
      }
    },
    "/v1/leaderboard/{leaderboardId}/participant/{participantId}/athlete": {
      "get": {
        "tags": [
          "Athlete"
        ],
        "operationId": "GetAthlete",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "leaderboardId",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "path",
            "name": "participantId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/DetailedAthleteResponse"
            }
          }
        }
      }
    },
    "/v1/auth/start": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "StartAuthentication",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/StartAuthenticationRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/StravaOauthResponse"
            }
          }
        }
      }
    },
    "/v1/auth/finish": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "operationId": "FinishAuthentication",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/FinishAuthenticationRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/v1/server/{serverId}/channel/{channelId}/text": {
      "post": {
        "tags": [
          "Discord"
        ],
        "operationId": "SendTextMessageToChannel",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "serverId",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "path",
            "name": "channelId",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SendTextMessageRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/v1/server/{serverId}/channel/{channelId}/embed": {
      "post": {
        "tags": [
          "Discord"
        ],
        "operationId": "SendEmbedMessageToChannel",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "serverId",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "path",
            "name": "channelId",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SendEmbedMessageRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/v1/server/{serverId}/role/assignments": {
      "delete": {
        "tags": [
          "Discord"
        ],
        "operationId": "RemoveAllInstancesOfRole",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "serverId",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/RemoveAllRoleAssignmentsRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Discord"
        ],
        "operationId": "GrantRoleAssignments",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "serverId",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/GrantRoleAssignmentsRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/v1/user/text": {
      "post": {
        "tags": [
          "Discord"
        ],
        "operationId": "SendDM",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SendTextMessageRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/v1/leaderboard/{leaderboardId}/participant": {
      "get": {
        "tags": [
          "Participant"
        ],
        "operationId": "GetAll",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "leaderboardId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Participant"
              }
            }
          }
        }
      }
    },
    "/v1/leaderboard/{leaderboardId}/participant/{participantId}": {
      "get": {
        "tags": [
          "Participant"
        ],
        "operationId": "Get",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "leaderboardId",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "path",
            "name": "participantId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Participant"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Participant"
        ],
        "operationId": "Delete",
        "parameters": [
          {
            "in": "path",
            "name": "leaderboardId",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "path",
            "name": "participantId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "ActivityType": {
      "enum": [
        "AlpineSki",
        "BackcountrySki",
        "Canoeing",
        "Crossfit",
        "EBikeRide",
        "Elliptical",
        "Golf",
        "Handcycle",
        "Hike",
        "IceSkate",
        "InlineSkate",
        "Kayaking",
        "Kitesurf",
        "NordicSki",
        "Ride",
        "RockClimbing",
        "RollerSki",
        "Rowing",
        "Run",
        "Sail",
        "Skateboard",
        "Snowboard",
        "Snowshoe",
        "Soccer",
        "StairStepper",
        "StandUpPaddling",
        "Surfing",
        "Swim",
        "Velomobile",
        "VirtualRide",
        "VirtualRun",
        "Walk",
        "WeightTraining",
        "Wheelchair",
        "Windsurf",
        "Workout",
        "Yoga"
      ],
      "type": "string"
    },
    "SummaryActivityResponse": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int64",
          "type": "integer"
        },
        "externalId": {
          "type": "string"
        },
        "uploadId": {
          "format": "int64",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "distance": {
          "format": "float",
          "type": "number"
        },
        "movingTime": {
          "format": "int32",
          "type": "integer"
        },
        "elapsedTime": {
          "format": "int32",
          "type": "integer"
        },
        "totalElevationGain": {
          "format": "float",
          "type": "number"
        },
        "elevHigh": {
          "format": "float",
          "type": "number"
        },
        "elevLow": {
          "format": "float",
          "type": "number"
        },
        "type": {
          "$ref": "#/definitions/ActivityType"
        },
        "startDate": {
          "format": "date-time",
          "type": "string"
        },
        "startDateLocal": {
          "format": "date-time",
          "type": "string"
        },
        "timezone": {
          "type": "string"
        },
        "achievementCount": {
          "format": "int32",
          "type": "integer"
        },
        "kudosCount": {
          "format": "int32",
          "type": "integer"
        },
        "commentCount": {
          "format": "int32",
          "type": "integer"
        },
        "athleteCount": {
          "format": "int32",
          "type": "integer"
        },
        "photoCount": {
          "format": "int32",
          "type": "integer"
        },
        "totalPhotoCount": {
          "format": "int32",
          "type": "integer"
        },
        "trainer": {
          "type": "boolean"
        },
        "commute": {
          "type": "boolean"
        },
        "manual": {
          "type": "boolean"
        },
        "private": {
          "type": "boolean"
        },
        "flagged": {
          "type": "boolean"
        },
        "workoutType": {
          "format": "int32",
          "type": "integer"
        },
        "uploadIdStr": {
          "type": "string"
        },
        "averageSpeed": {
          "format": "float",
          "type": "number"
        },
        "maxSpeed": {
          "format": "float",
          "type": "number"
        },
        "hasKudoed": {
          "type": "boolean"
        },
        "gearId": {
          "type": "string"
        },
        "kilojoules": {
          "format": "float",
          "type": "number"
        },
        "averageWatts": {
          "format": "float",
          "type": "number"
        },
        "deviceWatts": {
          "type": "boolean"
        },
        "maxWatts": {
          "format": "int32",
          "type": "integer"
        },
        "weightedAverageWatts": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "DetailedAthleteResponse": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int64",
          "type": "integer"
        },
        "resourceState": {
          "format": "int32",
          "type": "integer"
        },
        "firstname": {
          "type": "string"
        },
        "lastname": {
          "type": "string"
        },
        "profileMedium": {
          "type": "string"
        },
        "profile": {
          "type": "string"
        },
        "city": {
          "type": "string"
        },
        "state": {
          "type": "string"
        },
        "country": {
          "type": "string"
        },
        "premium": {
          "type": "boolean"
        },
        "summit": {
          "type": "boolean"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "updatedAt": {
          "format": "date-time",
          "type": "string"
        },
        "followerCount": {
          "format": "int32",
          "type": "integer"
        },
        "friendCount": {
          "format": "int32",
          "type": "integer"
        },
        "ftp": {
          "format": "int32",
          "type": "integer"
        },
        "weight": {
          "format": "float",
          "type": "number"
        }
      }
    },
    "StartAuthenticationRequest": {
      "type": "object",
      "properties": {
        "redirectUrl": {
          "format": "uri",
          "type": "string"
        }
      }
    },
    "StravaOauthResponse": {
      "type": "object",
      "properties": {
        "accessToken": {
          "type": "string"
        },
        "refreshToken": {
          "type": "string"
        }
      }
    },
    "FinishAuthenticationRequest": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string"
        },
        "participantId": {
          "format": "int64",
          "type": "integer"
        },
        "leaderboardId": {
          "format": "int64",
          "type": "integer"
        }
      }
    },
    "SendTextMessageRequest": {
      "type": "object",
      "properties": {
        "text": {
          "type": "string"
        }
      }
    },
    "FieldViewModel": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "value": {
          "type": "object"
        },
        "inline": {
          "type": "boolean"
        }
      }
    },
    "SendEmbedMessageRequest": {
      "type": "object",
      "properties": {
        "fields": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/FieldViewModel"
          }
        },
        "title": {
          "type": "string"
        }
      }
    },
    "RemoveAllRoleAssignmentsRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        }
      }
    },
    "GrantRoleAssignmentsRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "userIds": {
          "type": "array",
          "items": {
            "format": "int64",
            "type": "integer"
          }
        }
      }
    },
    "Participant": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int64",
          "type": "integer"
        },
        "stravaId": {
          "format": "int64",
          "type": "integer"
        },
        "leaderboardId": {
          "format": "int64",
          "type": "integer"
        }
      }
    }
  }
}